requirements.txt
python-dotenv
face_recognition
opencv-python
pyqt5
sounddevice
soundfile
flask
flask_restful
flask_sqlalchemy
simpleaudio
pyinstaller
requests
python-dateutil
mu-editor
pyqtgraph
